backend/requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
fastapi==0.85.0
gmqtt==0.6.11
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
paho-mqtt==1.6.1
psycopg2-binary==2.9.4
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.17.6
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
